+++ CHANGELOG.md
@@ -59,6 +59,10 @@ This is a major release, and it might be not compatible with your current usage 
     -   `toggler-caretleft`
     -   `toggler-micon`
     -   `toggler-micoff`
+-   `<Modal />`:
+    -   `preventReactFlowEvents`: Adds 'nopan', 'nowheel' and 'nodrag' classes to Modal's overlay classes in order to prevent react-flow to react to drag and pan actions in modals.
+        This is enabled by default for <SimpleDialog />.
+
 
 ### Removed
 
@@ -86,8 +90,6 @@ This is a major release, and it might be not compatible with your current usage 
 
 ### Fixed
 
--   `<Modal />`:
-    -   Add 'nopan', 'nowheel' and 'nodrag' classes to Modal's overlay classes in order to always prevent react-flow to react to drag and pan actions in modals.
 -   `<CodeAutocompleteField />`:
     -   In multiline mode, validation errors might be highlighted incorrectly (relative line offset added).